+++ ios/Flutter/flutter_export_environment.sh
@@ -1,12 +1,12 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=D:\DevTools\flutter"
-export "FLUTTER_APPLICATION_PATH=D:\solo\tftp"
+export "FLUTTER_ROOT=D:\DevTools\Android\flutter"
+export "FLUTTER_APPLICATION_PATH=E:\solo\tftp"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib\main.dart"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=1.2.0"
-export "FLUTTER_BUILD_NUMBER=1.2.0"
+export "FLUTTER_BUILD_NAME=1.2.3"
+export "FLUTTER_BUILD_NUMBER=1.2.3"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"